refactor: split completion provider and apply index.js import setting
